Move js-tree files out from /containers subfolder
