+++ chrome/content/zotero/itemTreeColumns.jsx
