+++ chrome/content/zotero/libraryTree.js
