+++ chrome/content/zotero/tools/build_typeSchemaData.html
@@ -7,7 +7,7 @@
 <p id="result"></p>
 <script src="../include.js"></script>
 <script type="text/javascript">
-	import FilePicker from 'zotero/filePicker';
+	import FilePicker from 'zotero/modules/filePicker';
 	
 	(async function () {
 		// Create schema
