+++ chrome/content/zotero/tools/csledit.js
@@ -23,7 +23,7 @@
     ***** END LICENSE BLOCK *****
 */
 
-import FilePicker from 'zotero/filePicker';
+import FilePicker from 'zotero/modules/filePicker';
 
 var Zotero_CSL_Editor = new function() {
 	this.init = init;
